+++ test/test_factions.rb
@@ -13,22 +13,23 @@ describe RuneterraCards do
   4	PZ	Piltover & Zaun
   5	SI	Shadow Isles
   6	BW	Bilgewater
+  9	MT	Mount Targon
   TABLE
 
-  FACTIONS = FACTION_TABLE.split("\n").map { |line| line.split[1] }.freeze
+  FACTIONS = FACTION_TABLE.split("\n").map(&:split).map{|line| [line[0].to_i, line[1], line[2]]}.freeze
 
   describe 'FACTION_IDENTIFIERS_FROM_INT' do
-    FACTIONS.each_with_index do |faction, index|
-      it "has #{faction} at position #{index}" do
-        _(RuneterraCards::FACTION_IDENTIFIERS_FROM_INT[index]).must_equal faction
+    FACTIONS.each do |integer_identifier, faction_identifier, faction_name|
+      it "#{faction_name}: integer identifier #{integer_identifier} => #{faction_identifier}" do
+        _(RuneterraCards::FACTION_IDENTIFIERS_FROM_INT[integer_identifier]).must_equal faction_identifier
       end
     end
   end
 
   describe 'FACTION_INTS_FROM_IDENTIFIER' do
-    FACTIONS.each_with_index do |faction, index|
-      it "gives #{index} when looking up '#{faction}'" do
-        _(RuneterraCards::FACTION_INTS_FROM_IDENTIFIER[faction]).must_equal index
+    FACTIONS.each do |integer_identifier, faction_identifier, faction_name|
+      it "#{faction_name}: faction identifier #{faction_identifier} => #{integer_identifier}" do
+        _(RuneterraCards::FACTION_INTS_FROM_IDENTIFIER[faction_identifier]).must_equal integer_identifier
       end
     end
   end